게시글 수정 시 권한 확인

+++ src/main/java/com/example/prj1be/controller/BoardController.java
@@ -56,7 +56,18 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id){
+    public ResponseEntity remove(@PathVariable Integer id,
+                                @SessionAttribute(value="login",required = false)Member login){
+
+        if (login==null){
+            return  ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id,login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+
         if (service.remove(id)){
             return ResponseEntity.ok().build();
         }else {
@@ -65,7 +76,17 @@ public class BoardController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Board board) {
+    public ResponseEntity edit(@RequestBody Board board,
+                               @SessionAttribute(value = "login" ,required = false) Member login ) {
+
+        if(login==null){
+            return  ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();// 401
+        }
+
+        if (!service.hasAccess(Integer.valueOf(board.getId()), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
 //        System.out.println("board = " + board);
       if (service.validate(board)){
         if ( service.update(board)){
